allow wicd daemon to start even if /etc/resolv.conf is a relative symlink

+++ wicd/wicd-daemon.py
@@ -1792,7 +1792,7 @@ def main(argv):
         backup_location = wpath.varlib + 'resolv.conf.orig'
         # don't back up if .orig exists, probably there cause
         # wicd exploded
-        if not os.path.exists(backup_location):
+        if not os.path.exists(backup_location) and not os.path.islink(backup_location):
             if os.path.islink('/etc/resolv.conf'):
                 dest = os.readlink('/etc/resolv.conf')
                 os.symlink(dest, backup_location)
